+++ src/types/product_type.jl
@@ -5,14 +5,14 @@ export Product
 export id
 
 # code
-struct Product{T <: Union{String, Int}}
+struct Product{T <: Union{String, Int64}}
 
     product_id::T
 
 end
 
 """
-    function Base.convert(::Type{Product}, product_id::T) where {T <: Union{AbstractString, Int}}
+    function Base.convert(::Type{Product}, product_id::T) where {T <: Union{AbstractString, Int64}}
 
 Single-line funcition to automatically convert permissible types to Product.
 Overload function in Base module.
@@ -21,7 +21,7 @@ first argument:    Product (user defined type)
 product_id:        ID to be converted to Product type
 """
 
-Base.convert(::Type{Product}, product_id::T) where {T <: Union{AbstractString, Int}} = Product(product_id)
+Base.convert(::Type{Product}, product_id::T) where {T <: Union{AbstractString, Int64}} = Product(product_id)
 
 """
     function id(obj::Product)
